add inference capabilities to conve, convkb, distmult, ntn, proje_pointwise, slm and sme
